build : middleware based on role and permission

+++ routes/web.php
@@ -31,40 +31,80 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-// Dashboard
+
+// DEFAULT DASHBOARD & PROFILE
 Route::middleware('auth')->group(function () {
     Route::get('/dashboard', [DashboardController::class, 'index'])->middleware(['auth'])->name('dashboard');
-    Route::resource('/users', UserController::class);
-    Route::resource('/customers', CustomerController::class);
-    Route::resource('/suppliers', SupplierController::class);
-    Route::resource('/employees', EmployeeController::class);
-    Route::resource('/advance-salary', AdvanceSalaryController::class)->except(['show']);
 
+    Route::get('/profile', [ProfileController::class, 'show'])->name('profile');
+    Route::get('/profile/edit', [ProfileController::class, 'edit'])->name('profile.edit');
+    Route::put('/profile', [ProfileController::class, 'update'])->name('profile.update');
+    Route::get('/profile/change-password', [ProfileController::class, 'changePassword'])->name('profile.change-password');
+});
+
+// ====== USERS ======
+Route::middleware(['permission:user.menu'])->group(function () {
+    Route::resource('/users', UserController::class)->except(['show']);
+});
+
+// ====== CUSTOMERS ======
+Route::middleware(['permission:customer.menu'])->group(function () {
+    Route::resource('/customers', CustomerController::class);
+});
+
+// ====== SUPPLIERS ======
+Route::middleware(['permission:supplier.menu'])->group(function () {
+    Route::resource('/suppliers', SupplierController::class);
+});
+
+// ====== EMPLOYEES ======
+Route::middleware(['permission:employee.menu'])->group(function () {
+    Route::resource('/employees', EmployeeController::class);
+});
+
+// ====== EMPLOYEE ATTENDENCE ======
+Route::middleware(['permission:attendence.menu'])->group(function () {
+    Route::resource('/employee/attendence', AttendenceController::class)->except(['show', 'update', 'destroy']);
+});
+
+// ====== SALARY EMPLOYEE ======
+Route::middleware(['permission:salary.menu'])->group(function () {
     // PaySalary
     Route::resource('/pay-salary', PaySalaryController::class)->except(['show', 'create', 'edit', 'update']);
     Route::get('/pay-salary/history', [PaySalaryController::class, 'payHistory'])->name('pay-salary.payHistory');
     Route::get('/pay-salary/history/{id}', [PaySalaryController::class, 'payHistoryDetail'])->name('pay-salary.payHistoryDetail');
     Route::get('/pay-salary/{id}', [PaySalaryController::class, 'paySalary'])->name('pay-salary.paySalary');
 
-    // Employee Attendence
-    Route::resource('/employee/attendence', AttendenceController::class)->except(['show', 'update', 'destroy']);
+    // Advance Salary
+    Route::resource('/advance-salary', AdvanceSalaryController::class)->except(['show']);
+});
 
-    // Products
+// ====== PRODUCTS ======
+Route::middleware(['permission:product.menu'])->group(function () {
     Route::get('/products/import', [ProductController::class, 'importView'])->name('products.importView');
     Route::post('/products/import', [ProductController::class, 'importStore'])->name('products.importStore');
     Route::get('/products/export', [ProductController::class, 'exportData'])->name('products.exportData');
     Route::resource('/products', ProductController::class);
-    Route::resource('/categories', CategoryController::class);
+});
 
-    // POS
-    Route::get('/pos', [PosController::class, 'index'])->name('pos.index');
+// ====== CATEGORY PRODUCTS ======
+Route::middleware(['permission:category.menu'])->group(function () {
+    Route::resource('/categories', CategoryController::class);
+});
+
+// ====== POS ======
+Route::middleware(['permission:pos.menu'])->group(function () {
+    Route::get('/pos', [PosController::class,'index'])->name('pos.index');
     Route::post('/pos/add-cart', [PosController::class, 'addCart'])->name('pos.addCart');
     Route::post('/pos/update-cart/{rowId}', [PosController::class, 'updateCart'])->name('pos.updateCart');
     Route::get('/pos/delete-cart/{rowId}', [PosController::class, 'deleteCart'])->name('pos.deleteCart');
     Route::get('/pos/all-item', [PosController::class, 'allItem'])->name('pos.allItem');
     Route::post('/pos/create-invoice', [PosController::class, 'createInvoice'])->name('pos.createInvoice');
     Route::post('/pos/print-invoice', [PosController::class, 'printInvoice'])->name('pos.printInvoice');
+});
 
+// ====== ORDERS ======
+Route::middleware(['permission:orders.menu'])->group(function () {
     // Create Order
     Route::post('/pos/order', [OrderController::class, 'orderStore'])->name('pos.orderStore');
     Route::get('/orders/pending', [OrderController::class, 'pendingOrders'])->name('order.pendingOrders');
@@ -77,38 +117,31 @@ Route::middleware('auth')->group(function () {
     Route::get('/stock', [OrderController::class, 'stockManage'])->name('order.stockManage');
 });
 
-Route::controller(RoleController::class)->middleware('auth')->group(function () {
+// ====== ROLE CONTROLLER ======
+Route::middleware(['permission:roles.menu'])->group(function () {
     // Permissions
-    Route::get('/permission', 'permissionIndex')->name('permission.index');
-    Route::get('/permission/create', 'permissionCreate')->name('permission.create');
-    Route::post('/permission', 'permissionStore')->name('permission.store');
-    Route::get('/permission/edit/{id}', 'permissionEdit')->name('permission.edit');
-    Route::put('/permission/{id}', 'permissionUpdate')->name('permission.update');
-    Route::delete('/permission/{id}', 'permissionDestroy')->name('permission.destroy');
+    Route::get('/permission', [RoleController::class, 'permissionIndex'])->name('permission.index');
+    Route::get('/permission/create', [RoleController::class, 'permissionCreate'])->name('permission.create');
+    Route::post('/permission', [RoleController::class, 'permissionStore'])->name('permission.store');
+    Route::get('/permission/edit/{id}', [RoleController::class, 'permissionEdit'])->name('permission.edit');
+    Route::put('/permission/{id}', [RoleController::class, 'permissionUpdate'])->name('permission.update');
+    Route::delete('/permission/{id}', [RoleController::class, 'permissionDestroy'])->name('permission.destroy');
 
     // Roles
-    Route::get('/role', 'roleIndex')->name('role.index');
-    Route::get('/role/create', 'roleCreate')->name('role.create');
-    Route::post('/role', 'roleStore')->name('role.store');
-    Route::get('/role/edit/{id}', 'roleEdit')->name('role.edit');
-    Route::put('/role/{id}', 'roleUpdate')->name('role.update');
-    Route::delete('/role/{id}', 'roleDestroy')->name('role.destroy');
+    Route::get('/role', [RoleController::class, 'roleIndex'])->name('role.index');
+    Route::get('/role/create', [RoleController::class, 'roleCreate'])->name('role.create');
+    Route::post('/role', [RoleController::class, 'roleStore'])->name('role.store');
+    Route::get('/role/edit/{id}', [RoleController::class, 'roleEdit'])->name('role.edit');
+    Route::put('/role/{id}', [RoleController::class, 'roleUpdate'])->name('role.update');
+    Route::delete('/role/{id}', [RoleController::class, 'roleDestroy'])->name('role.destroy');
 
     // Role Permissions
-    Route::get('/role/permission', 'rolePermissionIndex')->name('rolePermission.index');
-    Route::get('/role/permission/create', 'rolePermissionCreate')->name('rolePermission.create');
-    Route::post('/role/permission', 'rolePermissionStore')->name('rolePermission.store');
-    Route::get('/role/permission/{id}', 'rolePermissionEdit')->name('rolePermission.edit');
-    Route::put('/role/permission/{id}', 'rolePermissionUpdate')->name('rolePermission.update');
-    Route::delete('/role/permission/{id}', 'rolePermissionDestroy')->name('rolePermission.destroy');
-});
-
-// Profile
-Route::middleware('auth')->group(function () {
-    Route::get('/profile', [ProfileController::class, 'show'])->name('profile');
-    Route::get('/profile/edit', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::put('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::get('/profile/change-password', [ProfileController::class, 'changePassword'])->name('profile.change-password');
+    Route::get('/role/permission', [RoleController::class, 'rolePermissionIndex'])->name('rolePermission.index');
+    Route::get('/role/permission/create', [RoleController::class, 'rolePermissionCreate'])->name('rolePermission.create');
+    Route::post('/role/permission', [RoleController::class, 'rolePermissionStore'])->name('rolePermission.store');
+    Route::get('/role/permission/{id}', [RoleController::class, 'rolePermissionEdit'])->name('rolePermission.edit');
+    Route::put('/role/permission/{id}', [RoleController::class, 'rolePermissionUpdate'])->name('rolePermission.update');
+    Route::delete('/role/permission/{id}', [RoleController::class, 'rolePermissionDestroy'])->name('rolePermission.destroy');
 });
 
 require __DIR__.'/auth.php';
